requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-rest-knox==5.0.1
cryptography==43.0.0
cloudinary==1.41.0
requests==2.32.3